# Conflicts:
#	build.gradle
#	src/main/java/pickyeater/basics/mealplan/MealPlan.java
#	src/main/java/pickyeater/basics/mealplan/PickyMealPlan.java
#	src/main/java/pickyeater/basics/user/PickyUserStatus.java
#	src/main/java/pickyeater/basics/user/UserStatus.java
#	src/main/java/pickyeater/builders/NutrientsBuilder.java
#	src/main/java/pickyeater/builders/PickyUserBuilder.java
#	src/main/java/pickyeater/builders/UserBuilder.java
#	src/main/java/pickyeater/database/jsonutils/FoodCreator.java
#	src/main/java/pickyeater/database/jsonutils/JSONCreator.java

+++ src/main/java/pickyeater/basics/food/NutrientType.java
@@ -7,13 +7,13 @@ package pickyeater.basics.food;
 public enum NutrientType {  // TO FIX
     PROTEIN(4), CARBOHYDRATE(4), FAT(9), ALCOHOL(7);
 
-    double caloriesPerGram;
+    float caloriesPerGram;
 
-    NutrientType(double caloriesPerGram) {
+    NutrientType(float caloriesPerGram) {
         this.caloriesPerGram = caloriesPerGram;
     }
 
-    public double getCaloriesPerGram() {
+    public float getCaloriesPerGram() {
         return caloriesPerGram;
     }
 }
