+++ src/main/java/pickyeater/basics/user/PickyUserStatus.java
@@ -1,7 +1,6 @@
 package pickyeater.basics.user;
 
 import java.time.LocalDate;
-import java.util.Date;
 
 public class PickyUserStatus implements UserStatus {
 
@@ -12,7 +11,7 @@ public class PickyUserStatus implements UserStatus {
     private final Sex sex;
 
 
-    public PickyUserStatus(float weight, int height, float bodyFat, LocalDate dateOfBirth, Sex sex) {
+    public PickyUserStatus(float weight, int height,float bodyFat, LocalDate dateOfBirth, Sex sex) {
         this.weight = weight;
         this.bodyFat = bodyFat;
         this.height = height;
@@ -21,17 +20,17 @@ public class PickyUserStatus implements UserStatus {
     }
 
     @Override
-    public double getWeight() {
+    public float getWeight() {
         return weight;
     }
 
     @Override
-    public double getBodyFat() {
+    public float getBodyFat() {
         return bodyFat;
     }
 
     @Override
-    public double getHeight() {
+    public int getHeight() {
         return height;
     }
 
@@ -47,6 +46,12 @@ public class PickyUserStatus implements UserStatus {
 
     @Override
     public String toString() {
-        return "PickyUserStatus{" + "weight=" + weight + ", bodyFat=" + bodyFat + ", height=" + height + ", dateOfBirth=" + dateOfBirth + ", sex=" + sex + '}';
+        return "PickyUserStatus{" +
+                "weight=" + weight +
+                ", bodyFat=" + bodyFat +
+                ", height=" + height +
+                ", dateOfBirth=" + dateOfBirth +
+                ", sex=" + sex +
+                '}';
     }
 }
