+++ src/main/java/pickyeater/basics/user/UserStatus.java
@@ -1,13 +1,12 @@
 package pickyeater.basics.user;
 
 import java.time.LocalDate;
-import java.util.Date;
 
 public interface UserStatus {
 
-    double getWeight();
-    double getBodyFat();
-    double getHeight();
+    float getWeight();
+    float getBodyFat();
+    int getHeight();
     LocalDate getDateOfBirth();
     Sex getSex();
 }
